created RandomDistribution class to handle reading and writing to uncertml
