+++ lib9ml/python/nineml/abstraction_layer/expressions/__init__.py
@@ -1,1 +1,2 @@
-from .base import Alias, ODE, Expression, ExpressionWithSimpleLHS, RandomVariable
+from .base import Alias, ODE, Expression, ExpressionWithSimpleLHS
+from .random import RandomVariable, RandomDistribution
