+++ lib9ml/python/nineml/abstraction_layer/expressions/base.py
@@ -298,33 +298,3 @@ class Alias(BaseALObject, ExpressionWithSimpleLHS):
     def is_alias_str(cls, alias_str):
         """ Returns True if the string could be an alias"""
         return ':=' in alias_str
-
-
-class RandomVariable(BaseALObject):
-
-    element_name = 'RandomVariable'
-    defining_attributes = ('name', 'value', 'units')
-
-    def __init__(self, name, value, units):
-        self.name = name
-        self.value = value
-        self.units = units
-
-    def __repr__(self):
-        return ("RandomVariable(name={}, value={}, units={})"
-                .format(self.name, self.value, self.units))
-
-    def accept_visitor(self, visitor, **kwargs):
-        """ |VISITATION| """
-        return visitor.visit_randomvariable(self, **kwargs)
-
-    def name_transform_inplace(self, name_map):
-        try:
-            self.name = name_map[self.name]
-        except KeyError:
-            assert False, "'{}' was not found in name_map".format(self.name)
-
-    def set_units(self, units):
-        assert self.units == units, \
-            "Renaming units with ones that do not match"
-        self.units = units
